updating as an action
